refactor: Update users table migration to add is_admin field
Update the users table migration to add a new boolean field called is_admin. This field will be used to determine if a user is an admin or not. The default value for this field is set to false.

refactor: Update infaqs table migration foreign key constraint

Update the infaqs table migration to change the foreign key constraint for the distribusi_id column. The constraint is now referencing the distribusi_kenclengs table instead of the distribusis table.